SDL_SoundSystem test working, still unfinished business

+++ Minigin/Logging_Sound_System.h
@@ -16,7 +16,7 @@ namespace engine
 		Logging_Sound_System(engine::SoundSystem* ss) : _real_ss(ss) {};
 		~Logging_Sound_System() { delete _real_ss; }
 
-		void PlaySound(const engine::sound_id id, const float volume) override
+		void PlaySound(const engine::sound_id id, const int volume) override
 		{
 			_real_ss->PlaySound(id, volume);
 			std::cout << "Playing Sound " << id << " at volume " << volume << "\n";
